작성한 글이 있는 멤버 탈퇴

+++ src/main/java/com/caprj1/controller/MemberController.java
@@ -3,6 +3,7 @@ package com.caprj1.controller;
 import com.caprj1.domain.Member;
 import com.caprj1.service.MemberService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -41,8 +42,10 @@ public class MemberController {
     }
 
     @PostMapping("remove")
-    public String remove(Integer id) {
-        service.remove(id);
+    public String remove(Integer id, Authentication authentication) {
+        if (service.hasAccess(id, authentication)) {
+            service.remove(id);
+        }
         return "redirect:/member/signup";
     }
